feat(#143, #145): docs, move to pre package
