+++ models/model/pre/pipeline.py
@@ -20,12 +20,12 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from model.embeddings import Embeddings
-from model.pre_description import PreDescription
-from model.pre_name import PreName
-from model.pre_readme import PreReadme
-from model.pre_topics import PreTopics
-from model.vector import Vector
+from model.pre.embeddings import Embeddings
+from model.pre.pre_description import PreDescription
+from model.pre.pre_name import PreName
+from model.pre.pre_readme import PreReadme
+from model.pre.pre_topics import PreTopics
+from model.pre.vector import Vector
 
 """
 Pipeline.
@@ -49,10 +49,10 @@ class Pipeline:
         readme = PreReadme(self.repository["readme"]).tokens()
         description = PreDescription(self.repository["description"]).tokens()
         topics = PreTopics(self.repository["topics"]).tokens()
-        e_name = Embeddings(name, 30).embed()
-        e_readme = Embeddings(readme, 512).embed()
-        e_description = Embeddings(description, 100).embed()
-        e_topics = Embeddings(topics, 100).embed()
+        e_name = Embeddings(" ".join(name), 30).embed()
+        e_readme = Embeddings(" ".join(readme), 512).embed()
+        e_description = Embeddings(" ".join(description), 100).embed()
+        e_topics = Embeddings(" ".join(topics), 100).embed()
         return Vector(
             e_name,
             e_readme,
