+++ models/model_tests/pre/test_pre_name.py
@@ -21,35 +21,21 @@
 # SOFTWARE.
 import unittest
 
-from model.pre_readme import PreReadme
+from model.pre.pre_name import PreName
 
 """
-Test cases for PreReadme.
+Test cases for PreName.
 """
 
 
-class TestPreReadme(unittest.TestCase):
+class TestPreName(unittest.TestCase):
 
-    def test_preprocesses_readme_in_tokens(self):
-        tokens = PreReadme("""
-            ## Java Examples for Stream Processing with Apache Flink
-
-            This repository hosts Java code examples for
-            ["Stream Processing with Apache Flink"](//link).
-
-            **Note:** The Java examples are not complete yet. <br>
-            The [Scala examples](#scala) placed here.
-            """).tokens()
-        expected = [
-            "java", "examples", "stream", "process",
-            "apache", "flink", "repository", "host",
-            "java", "code", "examples", "stream",
-            "process", "apache", "flinklink", "note",
-            "java", "examples", "complete", "yet",
-            "scala", "examplesscala", "place"
-        ]
+    def test_preprocesses_name(self):
+        input = "streaming-with-flink/examples-java"
+        tokens = PreName(input).tokens()
+        expected = ["streaming", "flink", "example", "java"]
         self.assertEqual(
             tokens,
             expected,
-            f"received tokens {tokens} do not match with expected {expected}"
+            f"received tokens {tokens} for input: {input} do not match with expected {expected}"
         )
