+++ models/model_tests/pre/test_pre_readme.py
@@ -21,23 +21,35 @@
 # SOFTWARE.
 import unittest
 
-from model.embeddings import Embeddings
+from model.pre.pre_readme import PreReadme
 
 """
-Test cases for Embeddings. 
+Test cases for PreReadme.
 """
 
 
-class TestEmbeddings(unittest.TestCase):
+class TestPreReadme(unittest.TestCase):
 
-    def test_generates_embeddings_for_tokens(self):
-        shape = Embeddings(
-            ["apache", "kafka", "examples", "learning"],
-            4
-        ).embed().shape
-        expected = (4, 768)
+    def test_preprocesses_readme_in_tokens(self):
+        tokens = PreReadme("""
+            ## Java Examples for Stream Processing with Apache Flink
+
+            This repository hosts Java code examples for
+            ["Stream Processing with Apache Flink"](//link).
+
+            **Note:** The Java examples are not complete yet. <br>
+            The [Scala examples](#scala) placed here.
+            """).tokens()
+        expected = [
+            "java", "examples", "stream", "process",
+            "apache", "flink", "repository", "host",
+            "java", "code", "examples", "stream",
+            "process", "apache", "flinklink", "note",
+            "java", "examples", "complete", "yet",
+            "scala", "examplesscala", "place"
+        ]
         self.assertEqual(
-            shape,
+            tokens,
             expected,
-            f"received matrix's shape {shape} does not match with expected {expected}"
+            f"received tokens {tokens} do not match with expected {expected}"
         )
